+++ CHANGES.rst
@@ -14,7 +14,10 @@ Changes to API
 Bug Fixes
 ---------
 
-- 
+saturation
+~~~~~~~~~~
+
+- Adds a check for saturation bias in group 2 for nframes > 1. [#259]
 
 1.7.2 (2024-06-12)
 ==================